Move new view to frameworks
I'd previously put this in services.py, but that's more for actions on services, so this flow seems better placed
in frameworks.py
